Update software to fit v2 board

+++ src/motor.h
@@ -4,7 +4,7 @@
 #include "stdint.h"
 
 void motor_init(void);
-void motor_set_power_percentage(uint8_t);
+void motor_set_power_percentage(float);
 void motor_update_routine(void);
 
 #endif
